🐛 prisma dir bugfix
