+++ src/test/sdk/HttpError.ts
@@ -1,0 +1,1 @@
+export { HttpError } from "@nestia/fetcher";
