+++ src/test/sdk/IConnection.ts
@@ -1,0 +1,1 @@
+export type { IConnection } from "@nestia/fetcher";
