+++ src/test/sdk/Primitive.ts
@@ -1,0 +1,1 @@
+export type { Primitive } from "@nestia/fetcher";
