+++ src/test/sdk/functional/auth/index.ts
@@ -1,7 +1,7 @@
 /**
  * @packageDocumentation
  * @module api.functional.auth
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
 export * as oauth from "./oauth";
