+++ src/test/sdk/functional/auth/oauth/index.ts
@@ -1,7 +1,7 @@
 /**
  * @packageDocumentation
  * @module api.functional.auth.oauth
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
 import { Fetcher } from "@nestia/fetcher";
@@ -10,22 +10,18 @@ import typia from "typia";
 
 /**
  * 카카오 인증 로그인 페이지 주소 요청
- * 
+ *
  * @summary 카카오 인증 페이지 주소 요청
  * @tag authentication
  * @return 카카오 인증 페이지 주소
- * 
+ *
  * @controller AuthController.kakao()
  * @path GET /auth/oauth/kakao
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
-export async function kakao(
-    connection: IConnection,
-): Promise<kakao.Output> {
+export async function kakao(connection: IConnection): Promise<kakao.Output> {
     return !!connection.simulate
-        ? kakao.simulate(
-              connection,
-          )
+        ? kakao.simulate(connection)
         : Fetcher.fetch(
               connection,
               kakao.ENCRYPTED,
@@ -45,39 +41,35 @@ export namespace kakao {
 
     export const path = (): string => {
         return `/auth/oauth/kakao`;
-    }
+    };
     export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
         typia.random<Output>(g);
     export const simulate = async (
         connection: IConnection,
     ): Promise<Output> => {
         return random(
-            typeof connection.simulate === 'object' &&
+            typeof connection.simulate === "object" &&
                 connection.simulate !== null
                 ? connection.simulate
-                : undefined
+                : undefined,
         );
-    }
+    };
 }
 
 /**
  * 깃허브 인증 로그인 페이지 주소 요청
- * 
+ *
  * @summary 깃허브 인증 페이지 주소 요청
  * @tag authentication
  * @return 깃허브 인증 페이지 주소
- * 
+ *
  * @controller AuthController.github()
  * @path GET /auth/oauth/github
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
-export async function github(
-    connection: IConnection,
-): Promise<github.Output> {
+export async function github(connection: IConnection): Promise<github.Output> {
     return !!connection.simulate
-        ? github.simulate(
-              connection,
-          )
+        ? github.simulate(connection)
         : Fetcher.fetch(
               connection,
               github.ENCRYPTED,
@@ -97,17 +89,17 @@ export namespace github {
 
     export const path = (): string => {
         return `/auth/oauth/github`;
-    }
+    };
     export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
         typia.random<Output>(g);
     export const simulate = async (
         connection: IConnection,
     ): Promise<Output> => {
         return random(
-            typeof connection.simulate === 'object' &&
+            typeof connection.simulate === "object" &&
                 connection.simulate !== null
                 ? connection.simulate
-                : undefined
+                : undefined,
         );
-    }
+    };
 }
