+++ src/test/sdk/functional/index.ts
@@ -1,7 +1,7 @@
 /**
  * @packageDocumentation
  * @module api.functional
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
 export * as auth from "./auth";
