+++ src/test/sdk/functional/users/normals/index.ts
@@ -1,34 +1,25 @@
 /**
  * @packageDocumentation
  * @module api.functional.users.normals
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
 import { Fetcher } from "@nestia/fetcher";
 import type { IConnection } from "@nestia/fetcher";
 import typia from "typia";
 
+import type { INormal } from "./../../../../../app/user/normal/interface";
 import { NestiaSimulator } from "./../../../utils/NestiaSimulator";
-import type { INormal } from "./../../../../../src/app/user/normal/interface";
 
 /**
  * @controller UsersNormalsController.get()
  * @path GET /users/normals
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
-export async function get(
-    connection: IConnection,
-): Promise<get.Output> {
+export async function get(connection: IConnection): Promise<get.Output> {
     return !!connection.simulate
-        ? get.simulate(
-              connection,
-          )
-        : Fetcher.fetch(
-              connection,
-              get.ENCRYPTED,
-              get.METHOD,
-              get.path(),
-          );
+        ? get.simulate(connection)
+        : Fetcher.fetch(connection, get.ENCRYPTED, get.METHOD, get.path());
 }
 export namespace get {
     export type Output = Array<number>;
@@ -42,29 +33,29 @@ export namespace get {
 
     export const path = (): string => {
         return `/users/normals`;
-    }
+    };
     export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
         typia.random<Output>(g);
     export const simulate = async (
         connection: IConnection,
     ): Promise<Output> => {
         return random(
-            typeof connection.simulate === 'object' &&
+            typeof connection.simulate === "object" &&
                 connection.simulate !== null
                 ? connection.simulate
-                : undefined
+                : undefined,
         );
-    }
+    };
 }
 
 /**
  * 일반 사용자 공개 프로필 정보 요청 API
- * 
+ *
  * @summary 일반 사용자 공개 프로필 정보 요청
  * @tag normals
  * @param normal_id 사용자 id
  * @return 일반 사용자 공개 정보
- * 
+ *
  * @controller UsersNormalsController.geByNormalId()
  * @path GET /users/normals/:normalId
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
@@ -74,10 +65,7 @@ export async function geByNormalId(
     normal_id: string,
 ): Promise<geByNormalId.Output> {
     return !!connection.simulate
-        ? geByNormalId.simulate(
-              connection,
-              normal_id,
-          )
+        ? geByNormalId.simulate(connection, normal_id)
         : Fetcher.fetch(
               connection,
               geByNormalId.ENCRYPTED,
@@ -97,7 +85,7 @@ export namespace geByNormalId {
 
     export const path = (normal_id: string): string => {
         return `/users/normals/${encodeURIComponent(normal_id ?? "null")}`;
-    }
+    };
     export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
         typia.random<Output>(g);
     export const simulate = async (
@@ -107,14 +95,14 @@ export namespace geByNormalId {
         const assert = NestiaSimulator.assert({
             method: METHOD,
             host: connection.host,
-            path: path(normal_id)
+            path: path(normal_id),
         });
         assert.param("normalId")("string")(() => typia.assert(normal_id));
         return random(
-            typeof connection.simulate === 'object' &&
+            typeof connection.simulate === "object" &&
                 connection.simulate !== null
                 ? connection.simulate
-                : undefined
+                : undefined,
         );
-    }
+    };
 }
