+++ src/test/sdk/index.ts
@@ -1,0 +1,4 @@
+import * as api from "./module";
+
+export * from "./module";
+export default api;
