+++ src/test/sdk/module.ts
@@ -1,0 +1,5 @@
+export type * from "./IConnection";
+export type * from "./Primitive";
+export * from "./HttpError";
+
+export * as functional from "./functional";
