--- test/sdk/Primitive.ts
@@ -1,1 +1,0 @@
-export type { Primitive } from "@nestia/fetcher";
