--- test/sdk/index.ts
@@ -1,4 +1,0 @@
-import * as api from "./module";
-
-export * from "./module";
-export default api;
