--- test/sdk/module.ts
@@ -1,5 +1,0 @@
-export type * from "./IConnection";
-export type * from "./Primitive";
-export * from "./HttpError";
-
-export * as functional from "./functional";
