+++ tsconfig.build.json
@@ -3,5 +3,6 @@
     "compilerOptions": {
         "outDir": "./build" /* Specify an output folder for all emitted files. */
     },
-    "include": ["./src"]
+    "include": ["./src"],
+    "exclude": ["./src/test"]
 }
