+++ tsconfig.json
@@ -30,7 +30,7 @@
         "paths": {
             "@PRISMA": ["./db"],
             "@APP/*": ["./src/*"],
-            "@TEST/*": ["./test/*"]
+            "@TEST/*": ["./src/test/*"]
         } /* Specify a set of entries that re-map imports to additional lookup locations. */,
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
         // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
